+++ changelog.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.0.28
 - Feature: add deprecation checks for Liferay 2025.Q1 (7.4.3.132)
+- Feature: add deprecation check for LPD-8352
 - Feature: add basic support for Liferay 2026.Q1
 
 ## 0.0.27